main.py
from kivy.app import App
from kivy.uix.boxlayout import BoxLayout
from kivy.uix.camera import Camera
from kivy.uix.button import Button
from kivy.uix.label import Label
from kivy.graphics.texture import Texture
from kivy.clock import Clock
import cv2
import numpy as np
import tensorflow as tf
from PIL import Image
import time


class HelloWorld(BoxLayout):
    def capture_and_analyze(self):
        camera = self.ids['camera']
        timestr = time.strftime("%Y%m%d_%H%M%S")
        image_name = f"IMG_{timestr}.png"
        camera.export_to_png(image_name)
        self.analyze_image(image_name)

    def analyze_image(self, image_path):
        # This is a placeholder for your YOLO model inference
        # Replace this with your actual model loading and prediction logic
        try:
            # Load the TFLite model and allocate tensors.
            interpreter = tf.lite.Interpreter(model_path="your_model.tflite")
            interpreter.allocate_tensors()

            # Get input and output tensors.
            input_details = interpreter.get_input_details()
            output_details = interpreter.get_output_details()

            # Read and preprocess the image
            img = Image.open(image_path).resize((input_details[0]['shape'][1], input_details[0]['shape'][2]))
            input_data = np.expand_dims(img, axis=0)

            if input_details[0]['dtype'] == np.uint8:
                input_data = np.uint8(input_data)

            interpreter.set_tensor(input_details[0]['index'], input_data)
            interpreter.invoke()

            # The function `get_tensor()` returns a copy of the tensor data.
            # Use `tensor()` in order to get a pointer to the tensor.
            output_data = interpreter.get_tensor(output_details[0]['index'])
            
            # Process output_data to get your result
            # For demonstration, we'll just show a dummy result
            result_text = f"Analysis complete: {np.argmax(output_data)}"

        except Exception as e:
            result_text = f"Error: {e}. Please add your_model.tflite"

        self.ids.result_label.text = result_text


class HelloWorldApp(App):
    def build(self):
        return HelloWorld()


if __name__ == '__main__':
    HelloWorldApp().run()
