requirements.txt
Kivy==2.3.1
opencv-python-headless
tensorflow
Pillow
